Windows port: experimental dropout: fix narrowing casts
- dropout_program_factory.cpp: static_cast<uint8_t> the uint32_t
  cb_index passed into CBFormatDescriptor::buffer_index (uint8_t
  field) in a nested aggregate init (reported by MSVC as C2440 on
  the outer CBDescriptor aggregate); static_cast<uint32_t> the
  size_t grid.y value returned in the DropoutCoreSplit aggregate
  (C2397 narrowing).
- adaptive_pool target already compiled clean.

+++ ttnn/cpp/ttnn/operations/experimental/dropout/device/dropout_program_factory.cpp
@@ -70,7 +70,7 @@ inline void create_circular_buffer(
         .total_size = num_tiles * single_tile_size,
         .core_ranges = core_ranges,
         .format_descriptors = {{CBFormatDescriptor{
-            .buffer_index = cb_index,
+            .buffer_index = static_cast<uint8_t>(cb_index),
             .data_format = data_format,
             .page_size = single_tile_size,
         }}},
@@ -155,7 +155,7 @@ DropoutCoreSplit dropout_core_split(const Tensor& input) {
         tt::tt_metal::split_work_to_cores(grid, num_tiles);
     return {
         num_cores,
-        grid.y,
+        static_cast<uint32_t>(grid.y),
         all_cores,
         core_group_1,
         core_group_2,
